GateMP: Fix maxValueLen for GateMP's NameServer
Commit id 16c8220d5d895329de83b52060928968958f8a3d intruduced
a change to the max size of the value that GateMP will add
to its internal NameServer. However, the GateMP module's
setting for its NameServer's maxValueLen was not updated to
reflect the new max possible value for the case where
hostSupport is set to FALSE.

This patch updates the maxValueLen to be large enough
to accommodate the max value size used.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>